7. ** Find The Real Queen - Lab
Write a program that reads (8 x 8) matrix of characters from the console. The matrix represents a chess board with figures on it. The figures can be - queens as char 'q' or any other ASCII symbol. There will be more than one queen but only one queen will have valid position, which is not attacked from any other queen and does not attack any other queen. In other word in the way of the valid queen there are no other queens, but there may be any other ASCII symbol. Your task is to read the chessboard and find the position of the valid queen. According to chess rules the queen can attack all the cells in horizontal vertical and both diagonals which cross the queen position.
Input:
p a p a p q p a
p a p a p a p a
p a q a p a p a
p a p a p a p a
p a p q p a p a
p a p a p a p a
p a q a p a q a
p a p a p q p a